Massive reorganization into folders plus booleans for different optional versions
Added boolean switches to allow creating the text, a workbook and the
hints/solutions guide all
using the same source files.

Added boolean switches to allow for four versions: plain, S, N and SN.
The S flag means that logical
negation is denoted using $\sim$ rather than $\lnot$.  The N flag
distinguishes a version in which
the convention that $1$ is the smallest natural number is followed
throughout.  The absence
of the N flag indicates the original version of the text in which the
naturals do not contain $0$
in chapters 1 through 4 and (after the Peano axioms are introduced)
that convention is changed
to the more modern rule.

Reorganized source files into subdirectories for each chapter.

Lots of other fixes...